backend/requirements.txt
# FastAPI and related
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
slowapi==0.1.9

# Google Cloud services
google-cloud-firestore==2.20.1
google-cloud-storage==2.19.0
google-auth==2.36.0
firebase-admin==6.6.0

# VertexAI and ML
google-cloud-aiplatform==1.72.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client
httpx==0.28.1
requests==2.32.3
aiohttp==3.10.11

# Web scraping and HTML parsing
beautifulsoup4==4.12.3

# PDF generation
weasyprint==62.3
markdown==3.7

# Data processing
pandas==2.2.3
numpy==2.2.0

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Async support
# asyncio is built into Python 3.4+

# Date/time handling
python-dateutil==2.9.0

# Environment management
python-dotenv==1.0.1

# Logging
structlog==24.4.0
python-json-logger==2.0.7 